solved weird distance problem by not subtracting lookFrom from p

+++ test/camera.cpp
@@ -165,7 +165,7 @@ Scene::Scene(std::string file) {
       //camera lookfromx lookfromy lookfromz lookatx lookaty lookatz upx upy upz fov
       //  speciﬁes the camera in the standard way, as in homework 2.
       else if(!splitline[0].compare("camera")) {
-          int x, y, z;
+          float x, y, z;
         // lookfrom:
             x = atof(splitline[1].c_str());
             y = atof(splitline[2].c_str());
